+++ RoboMinerCpp/Robot.h
@@ -54,7 +54,7 @@ public:
     CRobot(int robotId, int maxTurns, int maxOre,
            int miningSpeed, int cpuSpeed,
            double forwardSpeed, double backwardSpeed, int rotateSpeed,
-           int robotSize);
+           double robotSize);
     virtual ~CRobot()                                           {}
 
     RobotAction getNextRobotAction();
