+++ RoboMinerCpp/robotcode/RobotProgram.cpp
@@ -45,7 +45,7 @@ CRobotProgram::CRobotProgram(int robotId,
                              int maxTurns, int maxOre,
                              int miningSpeed, int cpuSpeed,
                              double forwardSpeed, double backwardSpeed, int rotateSpeed,
-                             int robotSize) :
+                             double robotSize) :
     CRobot(robotId, maxTurns, maxOre,
            miningSpeed, cpuSpeed,
            forwardSpeed, backwardSpeed, rotateSpeed,
